test(frontend): extract app-core helpers + add unit tests
