+++ index.html
@@ -300,6 +300,7 @@
     </div>
 
     <script src="gateway-client.js"></script>
+    <script src="/lib/app-core.js"></script>
     <script src="app.js"></script>
   </body>
 </html>
